Add files via upload
Hopefully this is more readable. I seperated the files into two project. DataGenerator really just works with the data and can convert the Parsed.csv in the matrix that we need. I fixed the NMF and Wnnf files so that they use the DataGenerator as a dependency.

Should be understanable once you read the files :)